Door opening, closing; key using
Improved door draging script because in previous version we had 3 different scripts with wrong working and finnaly bad code.
Also added a keys for different locations on our map(including enum with places). So now player could open the door if door is not needed the key or vice versa if the door needs it player must first find the right key and press button to open it.